couple more explicit casts

+++ Server/RequestHandler.cpp
@@ -87,7 +87,7 @@ void RequestHandler::handle_request(const http::Request &req, http::Reply &reply
         {
             reply = http::Reply::StockReply(http::Reply::badRequest);
             reply.content.clear();
-            const int position = std::distance(request.begin(), iter);
+            const unsigned position = static_cast<unsigned>(std::distance(request.begin(), iter));
             JSON::Object json_result;
             json_result.values["status"] = 400;
             std::string message = "Query string malformed close to position ";
@@ -112,7 +112,7 @@ void RequestHandler::handle_request(const http::Request &req, http::Reply &reply
         }
 
         // set headers
-        reply.headers.emplace_back("Content-Length", UintToString(reply.content.size()));
+        reply.headers.emplace_back("Content-Length", UintToString(static_cast<unsigned>(reply.content.size())));
         if ("gpx" == route_parameters.output_format)
         {   // gpx file
             reply.headers.emplace_back("Content-Type", "application/gpx+xml; charset=UTF-8");
